LPD-82541 Move test to objectDefinition.spec.ts
