+++ modules/test/playwright/tests/object-web/main/objectCrud.spec.ts
@@ -3,10 +3,6 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {
-	ObjectDefinition,
-	ObjectRelationshipAPI,
-} from '@liferay/object-admin-rest-client-js';
 import {expect, mergeTests} from '@playwright/test';
 
 import {dataApiHelpersTest} from '../../../fixtures/dataApiHelpersTest';
@@ -14,9 +10,6 @@ import {featureFlagsTest} from '../../../fixtures/featureFlagsTest';
 import {isolatedSiteTest} from '../../../fixtures/isolatedSiteTest';
 import {loginTest} from '../../../fixtures/loginTest';
 import {objectPagesTest} from '../../../fixtures/objectPagesTest';
-import {getRandomInt} from '../../../utils/getRandomInt';
-import {waitForAlert} from '../../../utils/waitForAlert';
-import {localizationPagesTest} from '../../site-admin-web/main/fixtures/localizationPagesTest';
 import {generateObjectFields} from '../utils/generateObjectFields';
 
 const test = mergeTests(
@@ -25,7 +18,6 @@ const test = mergeTests(
 		'LPS-178052': {enabled: true},
 	}),
 	isolatedSiteTest,
-	localizationPagesTest,
 	loginTest(),
 	objectPagesTest
 );
@@ -71,152 +63,3 @@ test('Verify it is possible to view the custom object after restarting portal', 
 	).toBeVisible();
 });
 
-test.fixme(
-	'Verify it is possible to add an Object Entry Title Field when changing the localization on Instance Settings',
-	async ({
-		apiHelpers,
-		localizationInstanceSettingsPage,
-		page,
-		viewObjectDefinitionsPage,
-		viewObjectEntriesPage,
-	}) => {
-		let objectDefinition: ObjectDefinition;
-		let relationshipLabel: string;
-		let relationshipName: string;
-
-		await test.step('Create an Account and a Custom Object with a Relationship 1toM of Account to Custom Object', async () => {
-			await apiHelpers.headlessAdminUser.postAccount();
-
-			const objectFields = generateObjectFields({
-				objectFieldBusinessTypes: ['Text'],
-			});
-
-			objectDefinition =
-				await apiHelpers.objectAdmin.postRandomObjectDefinition({
-					objectFields,
-					status: {code: 0},
-				});
-
-			apiHelpers.data.push({
-				id: objectDefinition.id,
-				type: 'objectDefinition',
-			});
-
-			const objectRelationshipAPIClient =
-				await apiHelpers.buildRestClient(ObjectRelationshipAPI);
-
-			relationshipLabel = 'Relationship';
-			relationshipName = 'relationship' + getRandomInt();
-
-			await objectRelationshipAPIClient.postObjectDefinitionByExternalReferenceCodeObjectRelationship(
-				'L_ACCOUNT',
-				{
-					label: {en_US: relationshipLabel},
-					name: relationshipName,
-					objectDefinitionExternalReferenceCode2:
-						objectDefinition.externalReferenceCode,
-					objectDefinitionId2: objectDefinition.id,
-					objectDefinitionName2: objectDefinition.name,
-					type: 'oneToMany',
-				}
-			);
-		});
-
-		try {
-			await test.step('Change the default language in the virtual instance to Portuguese', async () => {
-				await localizationInstanceSettingsPage.goto('Language');
-
-				await localizationInstanceSettingsPage.defaultLanguageSelect.selectOption(
-					'pt_BR'
-				);
-
-				await localizationInstanceSettingsPage.saveSettings();
-			});
-
-			await test.step('Change the title field in the System Object and publish the Custom Object', async () => {
-				await viewObjectDefinitionsPage.goto();
-
-				await viewObjectDefinitionsPage.clickEditObjectDefinitionLink(
-					'Account'
-				);
-
-				await page
-					.getByRole('combobox', {name: 'Entry Title Field'})
-					.click();
-
-				await page.getByRole('option', {name: 'Type'}).click();
-
-				await viewObjectEntriesPage.saveObjectEntryButton.click();
-
-				await waitForAlert(page, 'The object was saved successfully.');
-
-				await page.waitForLoadState('networkidle');
-			});
-
-			await test.step('Create an entry using the title field selected', async () => {
-				await viewObjectEntriesPage.goto(objectDefinition.className);
-
-				await viewObjectEntriesPage.clickAddObjectEntry(
-					objectDefinition.label['en_US']
-				);
-
-				await page.getByLabel(relationshipLabel).click();
-
-				await page
-					.getByRole('option', {name: /business/i})
-					.first()
-					.click();
-
-				await viewObjectEntriesPage.saveObjectEntryButton.click();
-
-				await waitForAlert(page, 'The object was saved successfully.');
-
-				await page.waitForLoadState('networkidle');
-			});
-
-			await test.step('Verify the title field of System Object is present', async () => {
-				await viewObjectEntriesPage.backButton.click();
-
-				await expect(
-					page.getByRole('cell', {name: /business/i}).first()
-				).toBeVisible();
-
-				await page
-					.getByRole('cell', {name: /business/i})
-					.first()
-					.click();
-
-				await expect(page.getByText(/business/i).first()).toBeVisible();
-			});
-		}
-		finally {
-			await test.step('Restore default language to English', async () => {
-				await localizationInstanceSettingsPage.goto('Language');
-
-				await localizationInstanceSettingsPage.defaultLanguageSelect.selectOption(
-					'en_US'
-				);
-
-				await localizationInstanceSettingsPage.saveSettings();
-			});
-
-			await test.step('Restore Account Entry Title Field', async () => {
-				await viewObjectDefinitionsPage.goto();
-
-				await viewObjectDefinitionsPage.clickEditObjectDefinitionLink(
-					'Account'
-				);
-
-				await page
-					.getByRole('combobox', {name: 'Entry Title Field'})
-					.click();
-
-				await page.getByRole('option', {name: 'Name'}).click();
-
-				await viewObjectEntriesPage.saveObjectEntryButton.click();
-
-				await waitForAlert(page, 'The object was saved successfully.');
-			});
-		}
-	}
-);
